pyproject.toml
[project]
name = "potatomail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "appwrite>=8.0.0",
]
